Conversion des événements pour les éditeurs en entité

+++ core/src/main/java/fr/abes/licencesnationales/core/repository/EditeurEventRepository.java
@@ -1,9 +1,9 @@
 package fr.abes.licencesnationales.core.repository;
 
-import fr.abes.licencesnationales.core.entities.editeur.EditeurEventEntity;
+import fr.abes.licencesnationales.core.entities.editeur.event.EditeurEventEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EditeurEventRepository extends JpaRepository<EditeurEventEntity, Long> {
+public interface EditeurEventRepository extends JpaRepository<EditeurEventEntity, Integer> {
 }
